refactor: FE 리팩토링 백로그 정리
